travel_ai_backend/requirements.txt
# Django REST Framework Backend Requirements
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-environ==0.11.2
celery==5.3.4
redis==5.0.1

# Database
psycopg2-binary==2.9.9
pymongo==4.6.0
mongoengine==0.27.0

# AI/ML Stack
langchain==0.1.0
langchain-groq==0.1.0
langchain-openai==0.0.5
langchain-anthropic==0.1.0
openai==1.6.1
anthropic==0.8.1
groq==0.4.1

# Vector Databases
pinecone-client==2.2.4
weaviate-client==3.25.3
chromadb==0.4.18

# RAG & Embeddings
sentence-transformers==2.2.2
faiss-cpu==1.7.4
rank-bm25==0.2.2

# ML/DL
tensorflow==2.15.0
torch==2.1.2
scikit-learn==1.3.2
numpy==1.24.3
pandas==2.1.4

# API Documentation
drf-yasg==1.21.7
drf-spectacular==0.27.0

# Monitoring & Logging
sentry-sdk==1.39.2
prometheus-client==0.19.0

# Development
python-dotenv==1.0.0
pytest==7.4.3
pytest-django==4.7.0
black==23.11.0
flake8==6.1.0

# Production
gunicorn==21.2.0
whitenoise==6.6.0
